Bluetooth: Remember the state of 'Rename phone' activity on screen orientation change
Since we do not allow an empty name, when the Phone name is edited to zero
length, then the 'Rename' option in the activity will be grayed out. But,
when the orientation of the screen is changed with an empty phone name, the
option 'Rename' in the activity is enabled again. It is due the improper state
saved for the activity. Current change remembers the state properly and fixes
the issue.

CRs-fixed: 341486
(cherry picked from commit 025d1e4debbe86863850bb92c3db5506f3992903)

Change-Id: I3f6618cd63f4ddc0feaa60f4b95aa0b6b14ed44d