Force entire density syntax to complete before reboot fired
During instances where the first run was followed by selecting reboot
instead of "Got It", such as when the default density was already being
used and market was already configured, the code would be launched to
display the superuser prompt at the same time as the reboot command was
called. This method avoids this situation.