Make the volume overlay more functional 2/2
This patch is to add the settings in the Sound Menu.
And also to persist the users choice.

It depends on the patch: http://review.cyanogenmod.com/#change,14375

Under Volumes is a new List Preference for the volume overlay Style

Options are
   * Single
   * Expandable
   * Expanded
   * None

Changing this fires an intent that the Volume Panel is listening for.

Change-Id: I680124f534d667a0331cf0f0e3a5ec7eafb2068e